[skip ci] refactor
