+++ com.unity.renderstreaming/Runtime/Scripts/InputSystem/InputDeviceExtension.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b9703223ad9b06848b11ed494a9871a8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
